+++ extension.js
@@ -5,12 +5,11 @@ const vscode_1 = require("vscode");
 let terminal;
 function activate(context) {
     const cmds = [
-        ['language-j.createTerminal', createTerminal],
+        ['language-j.startTerminal', startTerminal],
         ['language-j.loadScript', loadScript],
         ['language-j.loadDisplayScript', loadDisplayScript],
-        ['language-j.executeSelection', executeSelection],
-        ['language-j.executeLine', executeLine],
-        ['language-j.executeLineAdvance', executeLineAdvance]
+        ['language-j.execute', execute],
+        ['language-j.executeAdvance', executeAdvance]
     ];
     for (const [n, f] of cmds) {
         vscode_1.commands.registerTextEditorCommand(n, f);
@@ -23,36 +22,79 @@ function deactivate(context) {
     }
 }
 exports.deactivate = deactivate;
+function createTerminal() {
+    const config = vscode_1.workspace.getConfiguration('j');
+    return vscode_1.window.createTerminal({
+        name: "Jconsole", shellPath: config.executablePath
+    });
+}
+vscode_1.window.onDidChangeActiveTerminal(nextTerminal => {
+    if (nextTerminal === undefined) {
+        return;
+    }
+    if (nextTerminal.name == "Jconsole") {
+        terminal = nextTerminal;
+    }
+    else {
+        const jTerminals = vscode_1.window.terminals.filter(t => t.name == "Jconsole");
+        terminal = jTerminals.length > 0 ? jTerminals[0] : null;
+    }
+});
+function getTerminal() {
+    if (terminal === null || terminal.exitStatus != undefined) {
+        terminal = createTerminal();
+    }
+    terminal.show(true);
+}
+function startTerminal() {
+    terminal = createTerminal();
+    terminal.show(false);
+}
 function loadScript(editor, _) {
-    createTerminal();
+    getTerminal();
     terminal.sendText(`load '${editor.document.fileName}'`);
 }
 function loadDisplayScript(editor, _) {
-    createTerminal();
+    getTerminal();
     terminal.sendText(`loadd '${editor.document.fileName}'`);
 }
-function executeSelection(editor, _) {
+function _execute(editor) {
+    let text = editor.document.getText(editor.selection);
+    let endPosition;
+    if (text.length === 0) {
+        endPosition = executeLine(editor);
+    }
+    else {
+        endPosition = executeSelection(editor);
+    }
+    return endPosition;
+}
+function execute(editor, _) {
+    _execute(editor);
+}
+function executeAdvance(editor, _) {
+    let endPosition = _execute(editor);
+    let offset = getNextNonBlankLineOffset(editor, endPosition);
+    vscode_1.commands.executeCommand('cursorMove', {
+        to: "down",
+        by: "wrappedLine",
+        value: offset
+    });
+    vscode_1.commands.executeCommand("cursorMove", {
+        to: "wrappedLineEnd"
+    });
+}
+function executeSelection(editor) {
+    getTerminal();
     const text = editor.document.getText(editor.selection);
     terminal.sendText(text, !text.endsWith('\n'));
+    return editor.selection.end;
 }
-function executeLine(editor, _) {
-    createTerminal();
-    const text = getExecutionText(editor);
-    console.log(text);
+function executeLine(editor) {
+    getTerminal();
+    const [text, endPosition] = getExecutionText(editor);
     terminal.sendText(text, !text.endsWith('\n'));
-}
-function executeLineAdvance(editor, edit) {
-    executeLine(editor, edit);
-    vscode_1.commands.executeCommand('cursorMove', { to: "down", by: "wrappedLine" });
-}
-function createTerminal() {
-    if (terminal == null || terminal.exitStatus != undefined) {
-        const config = vscode_1.workspace.getConfiguration('j');
-        terminal = vscode_1.window.createTerminal({
-            name: "Jconsole", shellPath: config.executablePath
-        });
-        terminal.show();
-    }
+    return endPosition;
 }
 function isMultilineStart(text) {
     const regex = /^.*\b([01234]|13|noun|adverb|conjunction|verb|monad|dyad)\s+(:\s*0|define)\b.*$/;
@@ -66,17 +108,24 @@ function getExecutionText(editor) {
     let lineIndex = editor.selection.active.line;
     let text = getLineText(editor, lineIndex);
     if (!isMultilineStart(text)) {
-        return text;
+        return [text, editor.selection.active];
     }
     while (lineIndex < editor.document.lineCount) {
         let nextLine = getLineText(editor, ++lineIndex);
         text += `\n${nextLine}`;
         if (isMultilineEnd(nextLine)) {
-            return text;
+            return [text, new vscode_1.Position(lineIndex, nextLine.length)];
         }
     }
     throw new Error("Incomplete multiline definition!");
 }
 function getLineText(editor, index) {
     return editor.document.lineAt(index).text;
+}
+function getNextNonBlankLineOffset(editor, endPosition) {
+    let lineIdx = 1 + endPosition.line;
+    while (lineIdx < editor.document.lineCount && getLineText(editor, lineIdx).trim().length === 0) {
+        lineIdx++;
+    }
+    return lineIdx - editor.selection.end.line;
 }